% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-summarize.R
\name{bootstrap_summarize}
\alias{bootstrap_summarize}
\alias{bootstrap_summarise}
\alias{bootstrap_do}
\alias{bootstrap}
\title{Bootstrapped actions on data frames}
\usage{
bootstrap_summarize(.data, ndraws, ..., .draw_key = ".draw")

bootstrap_summarise(.data, ndraws, ..., .draw_key = ".draw")

bootstrap_do(.data, ndraws, ..., .draw_key = ".draw")

bootstrap(.data, ndraws, .draw_key = ".draw")
}
\arguments{
\item{.data}{Data frame on which to operate.}

\item{ndraws}{Number of independent bootstrap draws to perform.}

\item{...}{Other arguments handed off to \code{[summarize()]} or \code{[do()]}.}

\item{.draw_key}{Name of the column that will hold an integer
running from 1 to \code{ndraws} indicating the bootstrap replicate.}
}
\description{
Perform bootstrapped actions on data frames (tibbles) respecting any
grouping that has been set up. The function \code{bootstrap_summarize()} is
a drop-in replacement for \code{[dplyr::summarize()]} and the function
\code{bootstrap_do()} is a drop-in replacement for \code{[dplyr::do()]}. The
function \code{bootstrap()} simply generates bootstrapped data sets and
returns them in a combined table.
}
