% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacao.R
\docType{data}
\name{cacao}
\alias{cacao}
\title{Ratings of chocolate bars}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1795 rows and 9 columns.}
\usage{
cacao
}
\description{
Ratings of almost 1800 bars of chocolate, collected by Brady Brelinski,
Founding Member of the Manhattan Chocolate Society.
}
\references{
Manhattan Chocolate Society, \url{http://flavorsofcacao.com}
}
\keyword{datasets}
