% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-outcomes.R
\name{sample_outcomes}
\alias{sample_outcomes}
\alias{confidence_band}
\title{Sample outcomes from a fitted model}
\usage{
sample_outcomes(model, newdata, times = 20, ...)

confidence_band(model, newdata, level = 0.95, n = 100, ...)
}
\description{
Currently, sampling of fitted models is only implemented for
GAMs fitted with \code{\link[mgcv:gam]{mgcv::gam()}}.
}
\examples{
library(mgcv)
fit <- gam(mpg ~ s(disp), data = mtcars, method = "REML")

newdata <- data.frame(
  disp = seq(
    min(mtcars$disp), max(mtcars$disp),
    length.out = 100
  )
)

sample_df <- sample_outcomes(fit, newdata, 10, unconditional = TRUE)
ggplot(mtcars, aes(disp, mpg)) +
 geom_point(color = "grey30", size = 0.5) +
 geom_line(data = sample_df, aes(group = .bootstrap), color = "#0072B2", size = 0.3)
}
