% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Create a sampler function useful for repeated reproducible sampling}
\usage{
sampler(times, size = 1, replace = FALSE, group = NULL,
  seed = NULL, key = ".draw", row = ".row", id = ".id")
}
\arguments{
\item{times}{Number of independent sampling draws to perform.}

\item{size}{Sample size for each random sample.}

\item{replace}{Bool indicating whether sampling should occur with or without replacement.}

\item{group}{An optional expression setting up the grouping to use for bootstrapping.
If not provided, any grouping present in the original dataset will be used.}

\item{seed}{Random seed to use.}

\item{key}{Name (as character) of the column that will hold an integer
running from 1 to \code{times} indicating the bootstrap replicate.}

\item{row}{Name (as character) of the column that will hold an integer
counting rows in the final bootstrapped dataset. Useful for
animations with gganimate.}

\item{id}{Name (as character) of the column that will hold an integer
running from 1 to n for each bootstrap, where n is the number
of observations in each group.}
}
\description{
Create a sampler function useful for repeated reproducible sampling
}
\examples{
spl <- sampler(3, 2)
spl(data.frame(letter = letters[1:4]))

library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point(aes(color = Species), alpha = 0.3) +
  geom_point(
    data = sampler(1, 5, group = Species),
    aes(fill = Species),
    color = "black", shape = 21
  ) +
  theme_bw()

# it is important to set grouping correctly for sampling
set.seed(1234)

df <- data.frame(
  type = c(rep("A", 100), rep("B", 10), rep("C", 3)),
  y = rnorm(113)
)

# incorrect: sampling ungrouped dataset leads to missing data
# in some categories
ggplot(df, aes(type, y)) +
  geom_pointrange(data = sampler(6, 3, replace = TRUE, seed = 559), stat = "summary") +
  facet_wrap(~.draw)

# correct: sampling within groups
ggplot(df, aes(type, y)) +
  geom_pointrange(data = sampler(6, 3, replace = TRUE, group = type, seed = 559), stat = "summary") +
  facet_wrap(~.draw)

# also correct: use grouped data frame
ggplot(group_by(df, type), aes(type, y)) +
  geom_pointrange(data = sampler(6, 3, replace = TRUE, seed = 559), stat = "summary") +
  facet_wrap(~.draw)

\dontrun{
library(gganimate)

p <- ggplot(iris, aes(Sepal.Length, Species, color = Species)) +
  geom_point(color = "grey50", alpha = 0.3, size = 2) +
  geom_point(data = sampler(20, 1, group = Species), size = 4) +
  scale_color_brewer(type = "qual", palette = 2, guide = "none") +
  theme_bw()

p + facet_wrap(~.draw)
p + transition_states(.draw, 1, 2)
}
}
\seealso{
\code{\link[=bootstrapper]{bootstrapper()}}
}
