% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-conf-strip.R
\docType{data}
\name{stat_conf_strip}
\alias{stat_conf_strip}
\alias{StatConfStrip}
\title{Horizontal confidence strips generated from statistical models}
\format{An object of class \code{StatConfStrip} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
stat_conf_strip(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", ..., confidence = 0.68, xlim = NULL,
  n = 501, na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE)
}
\arguments{
\item{confidence}{The confidence level used to calculate the \code{moe} statistic.
If the standard error is used as \code{moe}, then the confidence should be 0.68
(the default).}

\item{xlim}{Numeric vector of two numbers setting the range of x values to be
covered by the confidence strip. If not supplied, is taken from the x scale.}
}
\description{
The following aesthetics are understood by this stat (required aesthetics
are in bold):
\itemize{
\item \strong{\code{estimate}}: The estimate whose uncertainty is to be displayed
\item \strong{\code{moe}}: Margin of error
\item \code{confidence}: Confidence level of the moe, defaults to 0.68 (one standard deviation)
}
}
\keyword{datasets}
