% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summary-boot.R
\docType{data}
\name{stat_summary_boot}
\alias{stat_summary_boot}
\alias{StatSummaryBoot}
\title{Summarize data on bootstrap replicates}
\description{
Summarize data on bootstrap replicates
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(cyl, mpg, color = factor(stat(.draw)))) +
  stat_summary_boot(fun.y = "mean", geom = "point")

\dontrun{
library(gganimate)
ggplot(mtcars, aes(cyl, mpg)) +
  stat_summary_boot(
    aes(color = factor(stat(.draw))),
    fun.y = "mean", geom = "point", size = 2
  ) +
  transition_states(stat(.draw), 1, 1)

# doesn't work
ggplot(mtcars, aes(cyl, mpg)) +
  geom_point() +
  stat_summary_boot(
    aes(color = factor(stat(.draw))),
    fun.y = "mean", geom = "point", size = 2
  ) +
  transition_states(stat(.draw), 1, 1)
 }
}
\keyword{datasets}
