% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Create a bootstrapper function useful for repeated reproducible bootstrapping}
\usage{
bootstrapper(ndraws, ..., .seed = NULL, .draw_key = ".draw")
}
\arguments{
\item{ndraws}{Number of independent bootstrap draws to perform.}

\item{...}{Variables to group by.}

\item{.seed}{Random seed to use.}

\item{.draw_key}{Name of the column that will hold an integer
running from 1 to \code{ndraws} indicating the bootstrap replicate.}
}
\description{
Create a bootstrapper function useful for repeated reproducible bootstrapping
}
