% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapper.R
\name{bootstrapper}
\alias{bootstrapper}
\title{Create a bootstrapper function useful for repeated reproducible bootstrapping}
\usage{
bootstrapper(times, group = NULL, seed = NULL, key = ".bootstrap",
  row = ".row", id = ".id", original_id = ".original_id")
}
\arguments{
\item{times}{Number of independent bootstrap draws to perform.}

\item{group}{An optional expression setting up the grouping to use for bootstrapping.}

\item{seed}{Random seed to use.}

\item{key}{Name of the column that will hold an integer
running from 1 to \code{times} indicating the bootstrap replicate.}

\item{row}{Name of the column that will hold an integer
counting rows in the final bootstrapped dataset. Useful for
animations with gganimate.}

\item{id}{Name of the column that will hold an integer
running from 1 to n for each bootstrap, where n is the number
of observations in each group}

\item{original_id}{Name of the column that indicate the
original row that the bootstrapped row came from}
}
\description{
Create a bootstrapper function useful for repeated reproducible bootstrapping
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width)) +
  geom_point() +
  geom_smooth(data = bootstrapper(5, Species), aes(group = .bootstrap), se = FALSE) +
  facet_wrap(~Species)
}
