% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-outcomes.R
\name{sample_outcomes}
\alias{sample_outcomes}
\alias{confidence_band}
\title{Sample outcomes from a fitted model}
\usage{
sample_outcomes(model, newdata, times = 20, ...)

confidence_band(model, newdata, level = 0.95, n = 100, ...)
}
\description{
Currently, sampling of fitted models is only implemented for
GAMs fitted with \code{\link[mgcv:gam]{mgcv::gam()}}.
}
\examples{
library(mgcv)
library(ggplot2)
fit <- gam(mpg ~ s(disp), data = mtcars, method = "REML")

newdata <- data.frame(
  disp = seq(
    min(mtcars$disp), max(mtcars$disp),
    length.out = 100
  )
)

sample_df <- sample_outcomes(fit, newdata, 10, unconditional = TRUE)
conf <- confidence_band(fit, newdata, unconditional = TRUE)
ggplot(mtcars, aes(disp, mpg)) +
  geom_ribbon(data = conf, aes(ymin = lo, ymax = hi), fill="#80808080", color = NA) +
  geom_point(color = "grey30", size = 0.5) +
  geom_line(data = sample_df, aes(group = .bootstrap), color = "#0072B2", size = 0.3) +
  geom_line(data = conf, size = 1, color = "darkred")

\dontrun{
library(gganimate)

data(BlueJays, package = "Stat2Data")

fit <- gam(Skull ~ s(BillDepth) + s(BillWidth), data = BlueJays, method = "REML")

newdata <- expand.grid(
  BillWidth = seq(min(BlueJays$BillWidth), max(BlueJays$BillWidth), length.out = 20),
  BillDepth = seq(min(BlueJays$BillDepth), max(BlueJays$BillDepth), length.out = 20)
)

sample_df <- sample_outcomes(fit, newdata, 30, unconditional = TRUE)
ggplot(BlueJays, aes(BillDepth, BillWidth)) +
  geom_tile(data = sample_df, aes(fill = Skull), alpha = 0.5, color = NA) +
  geom_point(aes(size = Skull), color = "black", shape = 21, fill = NA) +
  scale_fill_viridis_c(option = "D", guide = guide_legend(override.aes = list(alpha = 0.5))) +
  theme_minimal() + #facet_wrap(~.bootstrap) # uncomment for static plot
  transition_states(.bootstrap, 1, 0)
}
}
