% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-confidence-density.R
\docType{data}
\name{stat_confidence_density}
\alias{stat_confidence_density}
\alias{StatConfidenceDensity}
\title{Confidence density distributions generated from estimate and margin of error}
\source{
Adrian W. Bowman. Graphs for Uncertainty. J. R. Statist. Soc. A 182:1-16, 2018.
\url{http://www.rss.org.uk/Images/PDF/events/2018/Bowman-5-Sept-2018.pdf}
}
\usage{
stat_confidence_density(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", ..., confidence = 0.95, xlim = NULL,
  n = 501, na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE)
}
\arguments{
\item{confidence}{The confidence level used to calculate the \code{moe} statistic.
This defaults to 0.95 (\code{moe} corresponds to 95\% confidence interval).}

\item{xlim}{Numeric vector of two numbers setting the range of x values to be
covered by the confidence density. If not supplied, is taken from the x scale.}
}
\description{
This stat generates normal densities from provided estimates plus margins
of error (at a specified confidence level). It can be used to estimate
the confidence density that underlies a given parameter estimate with
given margin of error.
}
\section{Details}{


The following aesthetics are understood by this stat (required aesthetics
are in bold):
\itemize{
\item \strong{\code{x}}: The estimate whose uncertainty is to be displayed
\item \strong{\code{moe}}: Margin of error
\item \code{confidence}: Confidence level used to calculate the \code{moe} statistic.
This defaults to 0.95 (\code{moe} corresponds to 95\% confidence interval).
}
}

\examples{
library(tidyverse)

cacao_small <- cacao \%>\%
  filter(location \%in\% c("Switzerland", "Canada", "U.S.A.", "Belgium"))

cacao_summary <- cacao_small \%>\%
  group_by(location) \%>\%
  summarize(
    sd = sd(rating),
    moe = sd*1.96,
    rating = mean(rating)
  )

ggplot(cacao_summary, aes(x = rating, y = location)) +
  stat_confidence_density(aes(moe = moe, fill = stat(ndensity)), height = 0.8) +
  geom_point(data = cacao_small, position = position_jitter(width = 0.05), size = 0.3) +
  geom_errorbarh(aes(xmin = rating - sd, xmax = rating + sd), height = 0.3, color = "darkred", size = 1) +
  geom_point(size = 3, color = "darkred") +
  theme_minimal()


library(ggridges)

cacao_se <- cacao_small \%>\%
  group_by(location) \%>\%
  summarize(
    se = sd(rating)/sqrt(n()),
    moe = se*1.96,
    rating = mean(rating)
  )

ggplot(cacao_se, aes(x = rating, y = location)) +
  stat_confidence_density(
    geom = "ridgeline",
    aes(moe = moe, height = stat(density)),
    alpha = NA, xlim = c(2.5, 3.75), scale = 0.08
  ) +
  theme_minimal()
}
\keyword{datasets}
