% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-confidence-density.R
\docType{data}
\name{stat_confidence_density}
\alias{stat_confidence_density}
\alias{StatConfidenceDensity}
\title{Confidence density distributions generated from estimate and margin of error}
\source{
Adrian W. Bowman. Graphs for Uncertainty. J. R. Statist. Soc. A 182:1-16, 2018.
\url{http://www.rss.org.uk/Images/PDF/events/2018/Bowman-5-Sept-2018.pdf}
}
\usage{
stat_confidence_density(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", ..., confidence = 0.95, xlim = NULL,
  n = 501, na.rm = FALSE, show.legend = FALSE, inherit.aes = TRUE)
}
\arguments{
\item{confidence}{The confidence level used to calculate the \code{moe} statistic.
This defaults to 0.95 (\code{moe} corresponds to 95\% confidence interval).}

\item{xlim}{Numeric vector of two numbers setting the range of x values to be
covered by the confidence density. If not supplied, is taken from the x scale.}
}
\description{
This stat generates normal densities from provided estimates plus margins
of error (at a specified confidence level). It can be used to estimate
the confidence density that underlies a given parameter estimate with
given margin of error.
}
\section{Details}{


The following aesthetics are understood by this stat (required aesthetics
are in bold):
\itemize{
\item \strong{\code{x}}: The estimate whose uncertainty is to be displayed
\item \strong{\code{moe}}: Margin of error
\item \code{confidence}: Confidence level used to calculate the \code{moe} statistic.
This defaults to 0.95 (\code{moe} corresponds to 95\% confidence interval).
}
}

\examples{
library(tidyverse)

params <- data.frame(
  group = letters[1:3],
  mean = c(1, 3, 2),
  sd = c(.8, .4, .7)
)
df_data <- mutate(
  params,
  value = purrr::map2(mean, sd, ~rnorm(250, .x, .y))
) \%>\% unnest()

df_summary <- group_by(df_data, group) \%>\%
  summarize(
    mean = mean(value),
    sd = sd(value),
    moe = sd*1.96
  )

ggplot(df_summary, aes(x = mean, y = group)) +
  stat_confidence_density(aes(moe = moe, fill = stat(ndensity)), height = 0.8, alpha = NA) +
  geom_point(data = df_data, aes(x = value), position = position_jitter(width = 0), size = 0.5) +
  geom_errorbarh(aes(xmin = mean - sd, xmax = mean + sd), height = 0.2, color = "darkred", size = 1) +
  geom_point(size = 3, color = "darkred") +
  scale_fill_gradient(low = "#132B4300", high = "#56B1F7FF") +
  theme_minimal()

library(ggridges)

ggplot(df_summary, aes(x = mean, y = group)) +
  stat_confidence_density(
    geom = "ridgeline",
    aes(moe = moe, height = stat(density)),
    alpha = NA, xlim = c(-1, 4)
  ) +
  theme_minimal()
}
\keyword{datasets}
